Added append_write
Added a way for the file open mode of the output file to be either append ('a') or write ('w'), depending on if the specified output file name already exists in the directory. 

Next, I want to create an option to to manually override the file open mode if the user wants to replace the existing file. There MUST be some kind of warning message/confirmation to make sure the user knows that the existing file will be overwritten. 

I would also like to make this program more user friendly with a list of options, confirmation that entered information is correct before proceeding, choosing which type of image file (banner/logo/colorfill/square) is to be analyzed, and some sort of formula for naming the output file.